Refactor: Change components extension
Changed all the components file extension from .js to .jsx
