+++ src/App.js
@@ -2,8 +2,7 @@ import './App.css';
 import Banner from './components/Banner';
 import Contact from './components/Contact';
 import Footer from './components/Footer';
-import Navbar from './components/NavBar';
-// import {NewsLetter} from './components/NewsLetter';
+import Navbar from './components/Navbar';
 import Projects from './components/Projects';
 import Skills from './components/Skills';
 import 'bootstrap/dist/css/bootstrap.min.css';
@@ -18,7 +17,6 @@ function App() {
       <Skills />
       <Projects />
       <Contact />
-      {/* <NewsLetter /> */}
       <Footer />
     </>
   );
